config_loader.py
import yaml

def load_config(config_file="config.yaml"):
    with open(config_file, "r") as f:
        config = yaml.safe_load(f)
    return config

def get_pdfs(config):
    return config.get("pdfs", [])
